Split canvas views into individual files
