+++ src/canvas/views/index.js
@@ -1,0 +1,6 @@
+// @flow
+
+export * from './FlamegraphView';
+export * from './ReactEventsView';
+export * from './ReactMeasuresView';
+export * from './TimeAxisMarkersView';
